requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.10
APScheduler==3.11.0
pytz==2025.2
qrcode==8.2
Pillow==11.3.0
requests==2.32.4
python-dotenv==1.1.1
Flask==3.1.1
gunicorn==23.0.0